Move LakeTablet to lake package
