+++ fe/fe-core/src/main/java/com/starrocks/persist/gson/GsonUtils.java
@@ -61,7 +61,6 @@ import com.starrocks.catalog.HiveResource;
 import com.starrocks.catalog.HudiResource;
 import com.starrocks.catalog.IcebergResource;
 import com.starrocks.catalog.JDBCResource;
-import com.starrocks.catalog.LakeTablet;
 import com.starrocks.catalog.LocalTablet;
 import com.starrocks.catalog.MapType;
 import com.starrocks.catalog.OdbcCatalogResource;
@@ -76,6 +75,7 @@ import com.starrocks.catalog.SinglePartitionInfo;
 import com.starrocks.catalog.SparkResource;
 import com.starrocks.catalog.StructType;
 import com.starrocks.catalog.Tablet;
+import com.starrocks.catalog.lake.LakeTablet;
 import com.starrocks.load.loadv2.LoadJob.LoadJobStateUpdateInfo;
 import com.starrocks.load.loadv2.SparkLoadJob.SparkLoadJobStateUpdateInfo;
 import com.starrocks.qe.SqlModeHelper;
