+++ fe/fe-core/src/main/java/com/starrocks/planner/MetaScanNode.java
@@ -4,7 +4,7 @@ package com.starrocks.planner;
 
 import com.google.common.collect.Lists;
 import com.starrocks.analysis.TupleDescriptor;
-import com.starrocks.catalog.LakeTablet;
+import com.starrocks.catalog.lake.LakeTablet;
 import com.starrocks.catalog.LocalTablet;
 import com.starrocks.catalog.MaterializedIndex;
 import com.starrocks.catalog.OlapTable;
