+++ fe/fe-core/src/test/java/com/starrocks/catalog/TabletStatMgrTest.java
@@ -4,6 +4,7 @@ package com.starrocks.catalog;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
+import com.starrocks.catalog.lake.LakeTablet;
 import com.starrocks.common.jmockit.Deencapsulation;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.thrift.TStorageMedium;
